CHORE: Group macros' CLI
