+++ examples/Makefile
@@ -6,10 +6,10 @@ install:
 	bundle exec pod install --verbose
 
 spm.fetch:
-	bundle exec pod spm fetch --all --verbose
+	bundle exec pod spm macro fetch --all --verbose
 
 spm.prebuild:
-	bundle exec pod spm prebuild --all --verbose
+	bundle exec pod spm macro prebuild --all --verbose
 
 build:
 	sh scripts/xcodebuild.sh
